+++ home/bin/src/update_ripgrep.sh
@@ -16,13 +16,17 @@ fi
 version="$(curl -s "https://github.com/BurntSushi/ripgrep/releases/latest" | grep -Eo "[0-9]+\.[0-9]+(\.[0-9]+)?")"
 
 if [ "$local_version" = "ripgrep ${version}" ]; then
-    echo "Already at latest version: ${local_version}"
+  echo "Already at latest version: ${local_version}"
 else
-    # get the package
-    file=ripgrep_${version}_${arch}.deb
-    curl -Ls -O "https://github.com/BurntSushi/ripgrep/releases/download/${version}/${file}"
-    # install it
+  # get the package
+  file=ripgrep_${version}_${arch}.deb
+  curl -Ls -O "https://github.com/BurntSushi/ripgrep/releases/download/${version}/${file}"
+  # install it
+  if [ -f $file ]; then
     ${SUDO} dpkg -i $file
     # remove the file
     rm -rf $file
+  else
+    echo "ERROR: File ${file} not found."
+  fi
 fi
